fix: fix the payout element to use fonts externally
